+++ src/game/mod.rs
@@ -2,10 +2,16 @@
 
 use bevy::prelude::*;
 
+mod animation;
 pub mod audio;
 mod movement;
 pub mod spawn;
 
 pub(super) fn plugin(app: &mut App) {
-    app.add_plugins((audio::plugin, movement::plugin, spawn::plugin));
+    app.add_plugins((
+        audio::plugin,
+        movement::plugin,
+        animation::plugin,
+        spawn::plugin,
+    ));
 }
